apps/common/app_requirements.txt
celery
redis
requests
flower
fastapi
uvicorn
gunicorn
gevent
openpyxl
pandas
xlrd
pymysql
python-dotenv
openai
aiohttp
httpx
mysql-connector-python
SQLAlchemy
mysqlclient